+++ API/include/dogbot/Servo.hh
@@ -126,13 +126,13 @@ namespace DogBotN {
     Json::Value ConfigAsJSON() const override;
 
     //! Get last reported state of the servo and the time it was taken.
-    bool GetState(TimePointT &tick,float &position,float &velocity,float &torque) const override;
+    bool GetState(TimePointT &tick,double &position,double &velocity,double &torque) const override;
 
     //! Estimate state at the given time.
     //! This will linearly extrapolate position, and assume velocity and torque are
     //! the same as the last reading.
     //! If the data is more than 5 ticks away from the
-    bool GetStateAt(TimePointT theTime,float &position,float &velocity,float &torque) const override;
+    bool GetStateAt(TimePointT theTime,double &position,double &velocity,double &torque) const override;
 
     //! Update torque for the servo.
     bool DemandTorque(float torque) override;
